do not store deployments as part of system.git repo by default as they are stored in db anyway which provides easier management and reduces need for zookeeper/helix for runtime cluster
(cherry picked from commit f1d02f496c56234584ac4990a8e442f82f5f69eb)

+++ jbpm-console-ng-business-domain/jbpm-console-ng-business-domain-backend/src/main/java/org/jbpm/console/ng/bd/backend/server/DeploymentManagerEntryPointImpl.java
@@ -57,6 +57,7 @@ public class DeploymentManagerEntryPointImpl implements DeploymentManagerEntryPo
   private static final Logger logger = LoggerFactory.getLogger(DeploymentManagerEntryPointImpl.class);
   private static final MergeMode defaultMergeMode = MergeMode.valueOf(System.getProperty("org.kie.dd.mergemode", MergeMode.MERGE_COLLECTIONS.toString()));
 
+  private boolean gitDeploymentsEnabled = Boolean.parseBoolean(System.getProperty("org.kie.git.deployments.enabled", "false"));
   private boolean autoDeployEnabled = Boolean.parseBoolean(System.getProperty("org.kie.auto.deploy.enabled", "true"));
 
   @Inject
@@ -396,9 +397,12 @@ public class DeploymentManagerEntryPointImpl implements DeploymentManagerEntryPo
    * runtime point of view
    */
   public void saveDeployment(@Observes @Deploy DeploymentEvent event) {
-    if (deploymentConfigService.getDeployment(event.getDeploymentId()) == null) {
-      deploymentConfigService.addDeployment(event.getDeploymentId(), event.getDeployedUnit().getDeploymentUnit());
-    }
+      if (!gitDeploymentsEnabled) {
+          return;
+      }
+      if (deploymentConfigService.getDeployment(event.getDeploymentId()) == null) {
+          deploymentConfigService.addDeployment(event.getDeploymentId(), event.getDeployedUnit().getDeploymentUnit());
+      }
   }
 
   /**
@@ -409,7 +413,10 @@ public class DeploymentManagerEntryPointImpl implements DeploymentManagerEntryPo
    * runtime point of view
    */
   public void removeDeployment(@Observes @Undeploy DeploymentEvent event) {
-    deploymentConfigService.removeDeployment(event.getDeploymentId());
+      if (!gitDeploymentsEnabled) {
+          return;
+      }
+      deploymentConfigService.removeDeployment(event.getDeploymentId());
   }
 
   /**
